kg_back/requirements.txt
fastapi >= 0.75.0
uvicorn >= 0.17.1
pdf2bib
python-multipart
black 
scapy 
scispacy
spacy-llm
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_lg-0.5.3.tar.gz
attrs
pdf2doi==1.5
pdfminer.six
nmslib==2.0.6
cattrs
sqlitedict
aiofiles
toml
protobuf==3.19.0
